Refactor App into smaller components
