+++ src/App.js
@@ -1,12 +1,9 @@
 import React, { useState, useEffect } from "react";
-import {
-  createAuth,
-  defaultConfig,
-  signInWithEmailAndPassword,
-  signInAnonymously,
-  onAuthStateChanged,
-} from "./firebase";
-import { signInWithCustomToken, signOut } from "firebase/auth";
+import { createAuth, defaultConfig, onAuthStateChanged } from "./firebase";
+import { signOut } from "firebase/auth";
+import SignInForms from "./components/SignInForms";
+import SavedClients from "./components/SavedClients";
+import UserPanel from "./components/UserPanel";
 
 // Initialize default Firebase auth
 const initialAuth = createAuth(defaultConfig);
@@ -14,16 +11,9 @@ const initialAuth = createAuth(defaultConfig);
 function App() {
   const [auth, setAuth] = useState(initialAuth);
   const [user, setUser] = useState(null);
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
   const [serverMessage, setServerMessage] = useState("");
-  const [cToken, setCToken] = useState("");
   const [fToken, setFToken] = useState(null);
   const [copied, setCopied] = useState(false);
-  const [clients, setClients] = useState([]);
-  const [newClientName, setNewClientName] = useState("");
-  const [newClientToken, setNewClientToken] = useState("");
-  const [newClientConfig, setNewClientConfig] = useState("");
 
   // Track Auth State
   useEffect(() => {
@@ -34,54 +24,7 @@ function App() {
     return () => unsubscribe();
   }, [auth]);
 
-  useEffect(() => {
-    const stored = localStorage.getItem("firebaseClients");
-    if (stored) {
-      try {
-        setClients(JSON.parse(stored));
-      } catch (_) {
-        setClients([]);
-      }
-    }
-  }, []);
 
-  const handleSignInWithCustomToken = async () => {
-    try {
-      // Example: Fetch the custom token from your backend
-      // const response = await fetch(
-      //   "http://localhost:3000/auth/get-custom-token",
-      //   {
-      //     method: "GET",
-      //     // or POST if needed
-      //   }
-      // );
-      // const data = await response.json();
-      // const { customToken } = data;
-
-      // 4. Sign in with the custom token
-      await signInWithCustomToken(auth, cToken);
-    } catch (error) {
-      console.error("SignInWithCustomToken Error:", error);
-    }
-  };
-
-  // Sign in with Email/Password
-  const handleEmailSignIn = async () => {
-    try {
-      await signInWithEmailAndPassword(auth, email, password);
-    } catch (error) {
-      console.error("Email Sign-In Error:", error);
-    }
-  };
-
-  // Sign in Anonymously
-  const handleAnonymousSignIn = async () => {
-    try {
-      await signInAnonymously(auth);
-    } catch (error) {
-      console.error("Anonymous Sign-In Error:", error);
-    }
-  };
 
   // Call the protected route
   const callProtectedEndpoint = async () => {
@@ -131,149 +74,25 @@ function App() {
     }
   };
 
-  const persistClients = (list) => {
-    setClients(list);
-    localStorage.setItem("firebaseClients", JSON.stringify(list));
-  };
-
-  const addClient = () => {
-    if (!newClientName || !newClientToken || !newClientConfig) return;
-    let configObj;
-    try {
-      configObj = JSON.parse(newClientConfig);
-    } catch (e) {
-      console.error("Invalid config JSON", e);
-      return;
-    }
-    const updated = [
-      ...clients,
-      { name: newClientName, token: newClientToken, config: configObj },
-    ];
-    persistClients(updated);
-    setNewClientName("");
-    setNewClientToken("");
-    setNewClientConfig("");
-  };
-
-  const removeClient = (index) => {
-    const updated = clients.filter((_, i) => i !== index);
-    persistClients(updated);
-  };
-
-  const signInClient = async (client) => {
-    try {
-      const newAuth = createAuth(client.config);
-      setAuth(newAuth);
-      await signInWithCustomToken(newAuth, client.token);
-    } catch (error) {
-      console.error("SignInWithCustomToken Error:", error);
-    }
-  };
 
   return (
     <div style={{ maxWidth: 600, margin: "0 auto" }}>
       <h1>Firebase Auth Demo</h1>
       {user ? (
-        <>
-          <p>Signed in as: {user.uid}</p>
-          <button onClick={callProtectedEndpoint}>
-            Call Protected Endpoint
-          </button>
-          {serverMessage && (
-            <pre style={{ background: "#f4f4f4", padding: "10px" }}>
-              {serverMessage}
-            </pre>
-          )}
-          <button onClick={handleShowFToken}>Show Token</button>
-          {fToken && (
-            <>
-              <pre style={{ background: "#f4f4f4", padding: "10px" }}>
-                {fToken}
-              </pre>
-              <button onClick={copyToClipboard}>Copy</button>
-            </>
-          )}
-          <button onClick={handleLogout} style={{ marginLeft: "1rem" }}>
-            Logout
-          </button>
-        </>
+        <UserPanel
+          user={user}
+          serverMessage={serverMessage}
+          fToken={fToken}
+          copied={copied}
+          onProtected={callProtectedEndpoint}
+          onShowToken={handleShowFToken}
+          onCopyToken={copyToClipboard}
+          onLogout={handleLogout}
+        />
       ) : (
         <>
-          <div>
-            <h2>Sign In with Email/Password</h2>
-            <input
-              placeholder="Email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              style={{ marginRight: 8 }}
-            />
-            <input
-              placeholder="Password"
-              type="password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              style={{ marginRight: 8 }}
-            />
-            <button onClick={handleEmailSignIn}>Sign In</button>
-          </div>
-          <div>
-            <h2>Or Sign In Anonymously</h2>
-            <button onClick={handleAnonymousSignIn}>Go Anonymous</button>
-          </div>
-          <div>
-            <h2>Sign In Custom Token</h2>
-            <input
-              placeholder="Customer token"
-              value={cToken}
-              onChange={(e) => setCToken(e.target.value)}
-              style={{ marginRight: 8 }}
-            />
-            <button onClick={handleSignInWithCustomToken}>
-              Go Custom Token
-            </button>
-          </div>
-          <div>
-            <h2>Saved Clients</h2>
-            {clients.map((c, idx) => (
-              <div key={idx} style={{ marginBottom: 8 }}>
-                <strong>{c.name}</strong>
-                <button
-                  onClick={() => signInClient(c)}
-                  style={{ marginLeft: 8 }}
-                >
-                  Sign In
-                </button>
-                <button
-                  onClick={() => removeClient(idx)}
-                  style={{ marginLeft: 4 }}
-                >
-                  Remove
-                </button>
-              </div>
-            ))}
-            <div>
-              <input
-                placeholder="Name"
-                value={newClientName}
-                onChange={(e) => setNewClientName(e.target.value)}
-                style={{ marginRight: 8 }}
-              />
-              <input
-                placeholder="Token"
-                value={newClientToken}
-                onChange={(e) => setNewClientToken(e.target.value)}
-                style={{ marginRight: 8 }}
-              />
-              <textarea
-                placeholder="Firebase config JSON"
-                value={newClientConfig}
-                onChange={(e) => setNewClientConfig(e.target.value)}
-                style={{ marginRight: 8, display: "block", width: "100%" }}
-                rows={3}
-              />
-              <button onClick={addClient}>Add</button>
-            </div>
-          </div>
+          <SignInForms auth={auth} />
+          <SavedClients setAuth={setAuth} />
         </>
       )}
     </div>
